+++ api/models/log.py
@@ -76,7 +76,7 @@ class Blocks:
         self.columns = [ 'challenge_id', 'plot_files', 'proofs_found', 'time_taken', 'farmed_block', 'created_at']
         self.rows = []
         plot_files = []
-        challenge_id = plots_past_filter = proofs_found = time_taken = farmed_block = None
+        challenge_id = plots_past_filter = proofs_found = time_taken = farmed_block = created_at = None
         cli_stdout.append('--') # add a trailing -- to force last parse
         for line in cli_stdout:
             try:
@@ -107,13 +107,13 @@ class Blocks:
                         plots_past_filter = proofs_found = time_taken = farmed_block = None
                         plot_files = []
                     else:
-                        app.logger.info("challenge_id: {0}".format(challenge_id))
-                        app.logger.info("plot_files: {0}".format(plot_files))
-                        app.logger.info("plots_past_filter: {0}".format(plots_past_filter))
-                        app.logger.info("proofs_found: {0}".format(proofs_found))
-                        app.logger.info("time_taken: {0}".format(time_taken))
-                        app.logger.info("created_at: {0}".format(created_at))
-                        app.logger.info("Missing farmed blocks data for farmed_block {0}".format(farmed_block))
+                        app.logger.debug("challenge_id: {0}".format(challenge_id))
+                        app.logger.debug("plot_files: {0}".format(plot_files))
+                        app.logger.debug("plots_past_filter: {0}".format(plots_past_filter))
+                        app.logger.debug("proofs_found: {0}".format(proofs_found))
+                        app.logger.debug("time_taken: {0}".format(time_taken))
+                        app.logger.debug("created_at: {0}".format(created_at))
+                        app.logger.debug("Missing farmed blocks data for farmed_block {0}".format(farmed_block))
             except:
                 app.logger.info("Failed to parse blocks line: {0}".format(line))
                 app.logger.info(traceback.format_exc())