+++ web/models/chia.py
@@ -372,7 +372,7 @@ class Wallets:
                 if re.match('^\s+-Type:\s+CAT$', line):
                     is_cat_wallet = True
             if is_cat_wallet:
-                app.logger.info("Ignoring balance of CAT type wallet named: {0}".format(lines[0][:-1]))
+                app.logger.debug("Ignoring balance of CAT type wallet named: {0}".format(lines[0][:-1]))
             else:
                 details.extend(chunk.split('\n'))
         return '\n'.join(details)
@@ -449,13 +449,32 @@ class Keys:
             except:
                 app.logger.info("Keys.init(): Unable to find a worker with hostname '{0}'".format(key.hostname))
                 displayname = key.hostname
+            parsed_details = key.details
+            try:
+                parsed_details = self.link_first_wallet_address(key.blockchain, key.details)
+            except:
+                traceback.print_exc()
+                parsed_details = key.details
             self.rows.append({ 
                 'displayname': displayname, 
                 'hostname': key.hostname,
                 'blockchain': key.blockchain,
                 'status': worker_status,
-                'details': key.details,
+                'details': parsed_details,
                 'updated_at': key.updated_at }) 
+    
+    def link_first_wallet_address(self, blockchain, details):
+        alltheblocks_blockchain = globals.get_alltheblocks_name(blockchain)
+        lines = []
+        for line in details.split('\n'):
+            if line.startswith('First wallet address'):
+                label = line.split(':')[0]
+                address = line.split(':')[1].strip()
+                link = "https://alltheblocks.net/{0}/address/{1}".format(alltheblocks_blockchain, address)
+                lines.append("{0}: <a target='_blank' class='text-white' href='{1}'>{2}</a>".format(label, link, address))
+            else:
+                lines.append(line)
+        return '\n'.join(lines)
 
 class Blockchains: